// frontend/src/shared/api/generated/model/completeStepResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * novalearn-api
 * OpenAPI spec version: 0.1.0
 */

export interface CompleteStepResponse {
  stars: number;
  study_star?: boolean;
  feynman_star?: boolean;
  test_star?: boolean;
  roadmap_progress?: number | null;
  mastery?: number | null;
  confidence?: number | null;
}