// frontend/src/shared/api/generated/model/roadmapOut.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * novalearn-api
 * OpenAPI spec version: 0.1.0
 */
import type { RoadmapSectionOut } from './roadmapSectionOut';

export interface RoadmapOut {
  folder_id: string;
  sections: RoadmapSectionOut[];
  total_lessons: number;
  completed_lessons: number;
  overall_progress: number;
}